fix(core): :recycle: create strategy to share ollama verification
